switched Demo2 to use win-kex made a backup of xrdp version

+++ Demo2.1-SetupKali.ps1
@@ -1,20 +1,22 @@
+# See https://www.linuxtechi.com/install-kali-linux-on-windows-wsl/
+
 # Check for Admin
 if ( -not(
   (New-Object Security.Principal.WindowsPrincipal (
-    [Security.Principal.WindowsIdentity]::GetCurrent())
-  ).IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator))){
-    Write-Warning 'Must be run as administrator'
-    return
-  }
+      [Security.Principal.WindowsIdentity]::GetCurrent())
+  ).IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator))) {
+  Write-Warning 'Must be run as administrator'
+  return
+}
 
-# Install Rrequired Windows Features
+# Install required Windows Features
+#region Steps 1 and 2
 $windowsFeatureList = @(
   'Microsoft-Windows-Subsystem-Linux'
   'VirtualMachinePlatform'
 )
-
 $installResult = Enable-WindowsOptionalFeature -Online -FeatureName $WindowsFeatureList -All -NoRestart
-if ($installResult.RestartNeeded){
+if ($installResult.RestartNeeded) {
   Write-Host 'The installation of Windows Features requires a reboot'
   $null = Read-Host 'Press Enter to reboot or Ctrl+C to cancel'
   Restart-Computer
@@ -32,19 +34,26 @@ try {
 
 # Set WSL2 as default
 $wslMajorVersion = ($wslVersion |
-  Select-String -Pattern 'wsl version: (\d)').Matches.Groups.Where({ $_.Name -eq 1 }).Value
+    Select-String -Pattern 'wsl version: (\d)').Matches.Groups.Where({ $_.Name -eq 1 }).Value
 Wait-Debugger
-if ($wslMajorVersion -ne 2){
+if ($wslMajorVersion -ne 2) {
   wsl --set-default-version 2
 }
+#endregion Steps 1 and 2
+
+# Step 3 is not needed if installing fresh per:
+# https://learn.microsoft.com/en-us/windows/wsl/install#upgrade-version-from-wsl-1-to-wsl-2
 
 # Install Kali Linux
-if ($wslList -match 'kali-linux'){
+#region Step 4
+if ($wslList -match 'kali-linux') {
   # already installed
-}else{
+} else {
   wsl --install -d kali-linux
 }
+#endregion Step 4
 
+#region extra setup
 # Copy setup script
 wsl -d kali-linux cp /mnt/c/Users/steve/OneDrive/Documents/Presentations/SafelyDetonate/setupkali.sh /home/steve/setupkali.sh
 
@@ -55,6 +64,14 @@ Write-Host './setuplinux.sh' -ForegroundColor Magenta
 Write-Host ("$([char]0x2191)" * 27) -ForegroundColor Yellow
 
 wsl -d kali-linux
+#endregion Setp 4
 
-# Back on Windows Host: RDP to local Kali instance
-mstsc.exe /v:127.0.0.1:3390 /prompt
+#region Step 5
+# Run Kali Linux and run Win-Kex script
+Write-Host ("$([char]0x2193)" * 27) -ForegroundColor Yellow
+Write-Host 'Run script: ' -ForegroundColor Cyan -NoNewline
+Write-Host './setupwinkex.sh' -ForegroundColor Magenta
+Write-Host ("$([char]0x2191)" * 27) -ForegroundColor Yellow
+
+# Run Kali using Win-Kex
+wsl -d kali-linux kex --win -s
